requirements.txt
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.7
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jsonpatch==1.33
jsonpath-python==1.1.5
jsonpointer==3.1.1
langchain==1.2.15
langchain-core==1.2.29
langchain-openrouter==0.2.1
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.13
langsmith==0.7.31
markdown-it-py==4.0.0
mdurl==0.1.2
openrouter==0.9.1
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.1
pydantic==2.13.1
pydantic_core==2.46.1
pyfiglet==1.0.4
Pygments==2.20.0
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.33.1
requests-toolbelt==1.0.0
rich==15.0.0
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
xxhash==3.6.0
zstandard==0.25.0
